Turn IPMB bridge into a RAII variable
IPMB bridge is temporary for one or more requests and only when it's
needed depending on the destination address. Before 2 functions were used
to set/clear the bridge but there were cases where an exception could
leave the bridge active, so the successive requests would fail. This
patch introduces the IpmbBridgeScoped() auto-managing class.